HEADER_FILE

+++ main.h
@@ -14,8 +14,8 @@ int	_printf(const char *format, ...);
 void	print_binary(unsigned int num, int *len);
 void	*_realloc(void *ptr, unsigned int old_size, unsigned int new_size);
 void	rev_string(char *s);
-void	_hexalower(unsigned int num, int *len);
-void	_hexaupper(unsigned int num, int *len);
+void	_hexalower(unsigned long num, int *len);
+void	_hexaupper(unsigned long num, int *len);
 void	_unsigned(unsigned long n, int *len);
 
 #endif
